added all file support and a data-flow is a img or other file
